refactor(tool): update command `v` to run script directly
