+++ package.json
@@ -34,7 +34,7 @@
     "dev:target": "yarn tools --watch --target",
     "apps": "yarn tools --dev --apps",
     "watch": "yarn tools --watch",
-    "v": "yarn tools --v",
+    "v": "node ./tools/scripts/get-npm-version.mjs",
     "bump": "yarn tools --v --patch",
     "bump:minor": "yarn tools --v --minor",
     "bump:major": "yarn tools --v --major",
